Move GetUserInfo function to seperate js file from Main.js
